requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
tensorflow==2.12.0
numpy==1.23.5
pillow==9.5.0